Channel-based client
This is a giant rework of client architecture.
It aims to fix all those horrible concurrency issues(multiple
goroutins call Read() multiple goroutines call Write(), all of which
have shared variables and any can trigger reconnect)
Now, there are 2 goroutines for each flow that push reads and writes.

+++ tapdance/flow.go
@@ -18,7 +18,7 @@ const (
 // Connection-oriented state
 type TapDanceFlow struct {
 	// tunnel index and start time
-	id           uint
+	id           uint64
 	startMs      uint64 // TODO: unused
 
 	// reference to global proxy
@@ -29,7 +29,7 @@ type TapDanceFlow struct {
 };
 
 // constructor
-func NewTapDanceFlow(proxy *TapdanceProxy, id uint) *TapDanceFlow {
+func NewTapDanceFlow(proxy *TapdanceProxy, id uint64) *TapDanceFlow {
 	state := new(TapDanceFlow)
 
 	state.proxy = proxy
